refactor template (flatten field)
